+++ src/frontend/src/components/Banner/Banner.jsx
@@ -1,11 +1,10 @@
 import React, { useEffect, useState } from 'react';
 import PropTypes from 'prop-types';
-import { Link } from 'gatsby';
 import { makeStyles } from '@material-ui/core/styles';
 import KeyboardArrowDownIcon from '@material-ui/icons/KeyboardArrowDown';
 import { Fab, Grid, Typography } from '@material-ui/core';
 import useSiteMetadata from '../../hooks/use-site-metadata';
-import DynamicBackgroundContainer from '../DynamicBackgroundContainer.jsx';
+import BannerDynamicText from '../BannerDynamicText/BannerDynamicText.jsx';
 
 const useStyles = makeStyles((theme) => ({
   h1: {
@@ -34,32 +33,6 @@ const useStyles = makeStyles((theme) => ({
   heroBanner: {
     height: '100vh',
   },
-  stats: {
-    position: 'absolute',
-    color: theme.palette.text.primary,
-    fontFamily: 'Roboto',
-    fontSize: '2rem',
-    display: 'block',
-    bottom: theme.spacing(12),
-    left: theme.spacing(8),
-    right: theme.spacing(8),
-    lineHeight: 'inherit',
-    letterSpacing: 'inherit',
-    transition: 'all linear 1s',
-    [theme.breakpoints.between('xs', 'sm')]: {
-      textAlign: 'left',
-      fontSize: '2rem',
-      left: theme.spacing(4),
-      right: theme.spacing(4),
-    },
-    [theme.breakpoints.between('md', 'lg')]: {
-      fontSize: '4rem',
-    },
-    [theme.breakpoints.up('xl')]: {
-      fontSize: '8rem',
-    },
-  },
-
   version: {
     position: 'absolute',
     opacity: 0.85,
@@ -95,15 +68,6 @@ const useStyles = makeStyles((theme) => ({
       bottom: theme.spacing(18),
     },
   },
-  addYours: {
-    color: theme.palette.text.primary,
-    textDecorationLine: 'underline',
-  },
-  dynamic: {
-    transition: 'opacity 1s ease-in-out',
-    backgroundColor: theme.palette.primary.main,
-    opacity: 0.9,
-  },
 }));
 
 function ScrollDown(props) {
@@ -120,22 +84,6 @@ function ScrollDown(props) {
   return (
     <div onClick={handleClick} role="presentation">
       {children}
-    </div>
-  );
-}
-
-function RetrieveBannerDynamicAssets() {
-  const classes = useStyles();
-
-  return (
-    <div className={classes.dynamic}>
-      <DynamicBackgroundContainer />
-      <Typography variant="caption" className={classes.stats}>
-        This year 83 of us have written over 250K words and counting.{' '}
-        <Link className={classes.addYours} to="/myfeeds">
-          Add yours!
-        </Link>
-      </Typography>
     </div>
   );
 }
@@ -172,7 +120,7 @@ export default function Banner() {
   return (
     <>
       <div className={classes.heroBanner}>
-        <RetrieveBannerDynamicAssets />
+        <BannerDynamicText />
 
         <Typography variant="h1" className={classes.h1}>
           {'Telescope'}
