+++ src/frontend/src/components/BannerImage/BannerImage.jsx
@@ -26,7 +26,7 @@ const useStyles = makeStyles((theme) => ({
 
 const randomGenerator = (min, max) => Math.floor(Math.random() * (max - min + 1)) + min;
 
-export default function DynamicBackgroundContainer() {
+export default function BannerImage() {
   const classes = useStyles();
 
   return (
